requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pymongo>=4.6.0
dnspython>=2.6.1
langchain>=0.2.15
langgraph>=0.2.22
langchain-openai>=0.2.0
pydantic>=2.7.0
python-dotenv>=1.0.1
google-generativeai>=0.7.2
gunicorn>=21.2.0
